+++ test/integration/commands/notify.spec.ts
@@ -1,24 +1,21 @@
 import {
+  NotifyEventType,
+  UpdateEventType
+} from '../../../src/modules/pubsub/types.js'
+import { clearDb } from '../../test-utils/clear.js'
+import { setupDatabase } from '../../test-utils/setup-db.js'
+import { disconnectDatabase } from '../../test-utils/disconnect.js'
+import {
   notify,
   pubSubConnectedToStreamerEvent,
   updateNotification,
-  updateTopicTypeForChannel,
   userNotificationIsExisting,
-  userIsAlreadyNotified
-} from '../../../commands/cmd/notify.js'
-import {
-  NotifyEventType,
-  UpdateEventType
-} from '../../../modules/pubsub/types.js'
-import {
-  getExampleNotificationEntity,
-  getExampleTwitchUserState,
-  getTwitchUserEntity
-} from '../../../spec/examples/user.js'
-import { clearDb } from '../../test-utils/clear.js'
-import { Notification, TwitchUser } from '../../../db/export-entities.js'
-import { setupDatabase } from '../../test-utils/setup-db.js'
-import { disconnectDatabase } from '../../test-utils/disconnect.js'
+  userIsAlreadyNotified,
+  updateTopicTypeForChannel
+} from '../../../src/commands/cmd/notify.js'
+import { TwitchUser } from '../../../src/db/export-entities.js'
+import { Notification } from '../../../src/db/export-entities.js'
+import { getExampleNotificationEntity, getExampleTwitchUserEntity, getExampleTwitchUserState } from '../../test-utils/example.js'
 
 describe('test notify command: ', () => {
   jasmine.DEFAULT_TIMEOUT_INTERVAL = 15000
@@ -34,8 +31,8 @@ describe('test notify command: ', () => {
   beforeEach(async () => {
     channel = 'testChannel'
     streamer = 'streamer'
-    notification = getExampleNotificationEntity()
-    user = getTwitchUserEntity()
+    notification = getExampleNotificationEntity({})
+    user = getExampleTwitchUserEntity({})
     jasmine.DEFAULT_TIMEOUT_INTERVAL = 15000
     await clearDb(hb.db.dataSource)
   })
@@ -139,7 +136,7 @@ describe('test notify command: ', () => {
 
       const result = await findNotification(user.id, streamer)
 
-      expect(result.live).toBeTruthy()
+      expect(result!.live).toBeTruthy()
     })
 
     it('update new db entry for user updates the notification', async () => {
@@ -151,7 +148,7 @@ describe('test notify command: ', () => {
 
       const result = await findNotification(notification.user.id, streamer)
 
-      expect(result.live).toBeTruthy()
+      expect(result!.live).toBeTruthy()
     })
   })
 
@@ -240,7 +237,7 @@ describe('test notify command: ', () => {
         name: streamer
       })
 
-      expect(updatedEntity.status).toBeTruthy()
+      expect(updatedEntity!.status).toBeTruthy()
     })
 
     it('should update setting for setting type', async () => {
@@ -258,7 +255,7 @@ describe('test notify command: ', () => {
         name: streamer
       })
 
-      expect(updatedEntity.setting).toBeTruthy()
+      expect(updatedEntity!.setting).toBeTruthy()
     })
     it('should create new entry if not existing with status type true', async () => {
       const id = 1
@@ -269,7 +266,26 @@ describe('test notify command: ', () => {
       })
 
       expect(createdEntity).not.toBeNull()
-      expect(createdEntity.status).toBeTruthy()
+      expect(createdEntity!.status).toBeTruthy()
+    })
+
+    it('creating new connection and invoking listen to topic function', async () => {
+      const message = [streamer, UpdateEventType.LIVE]
+      const userState = getExampleTwitchUserState({})
+      const returnedStreamerId = 1
+      await hb.db.userRepo.save(user)
+
+      spyOn(hb.api.twitch, 'getUserIdByName').and.resolveTo(returnedStreamerId)
+      spyOn(hb.pubSub, 'listenToTopic')
+
+      await notify.execute(channel, userState, message)
+      const expectedStreamerId = returnedStreamerId
+      const expectedNotifyType = NotifyEventType.STATUS
+
+      expect(hb.pubSub.listenToTopic).toHaveBeenCalledWith(
+        expectedStreamerId,
+        expectedNotifyType
+      )
     })
   })
 })
@@ -277,7 +293,7 @@ describe('test notify command: ', () => {
 async function findNotification(
   userId: number,
   streamer: string
-): Promise<Notification> {
+): Promise<Notification | null> {
   return hb.db.notificationRepo.findOne({
     where: {
       user: {